refactor(user): edit the Component name of layouts
